+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
+- [!202][!202] Re-add missing requirement for the first argument of a pipeline to be a clock
 - [!205][!205] Fix panic on method calls in let bindings
 - [!206][!206] Re-add working VCD translation. It now also translates more values
 
@@ -29,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Internal
 
 
+[!202]: https://gitlab.com/spade-lang/spade/-/merge_requests/202
 [!205]: https://gitlab.com/spade-lang/spade/-/merge_requests/205
 [!206]: https://gitlab.com/spade-lang/spade/-/merge_requests/206
 [!207]: https://gitlab.com/spade-lang/spade/-/merge_requests/207